// src/lib/api/github-files-client/src/models/DirectoryEntryDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mini-GitHub Files API
 * Servicio para contenido de archivos, commits, descargas y comparacion de branches.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GitObjectType } from './GitObjectType';
import {
    GitObjectTypeFromJSON,
    GitObjectTypeFromJSONTyped,
    GitObjectTypeToJSON,
} from './GitObjectType';

/**
 * Entrada de directorio (para listar contenido de carpetas)
 * @export
 * @interface DirectoryEntryDTO
 */
export interface DirectoryEntryDTO {
    /**
     * 
     * @type {string}
     * @memberof DirectoryEntryDTO
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof DirectoryEntryDTO
     */
    path: string;
    /**
     * 
     * @type {string}
     * @memberof DirectoryEntryDTO
     */
    sha: string;
    /**
     * 
     * @type {GitObjectType}
     * @memberof DirectoryEntryDTO
     */
    type: GitObjectType;
    /**
     * 
     * @type {number}
     * @memberof DirectoryEntryDTO
     */
    size?: number;
    /**
     * 
     * @type {string}
     * @memberof DirectoryEntryDTO
     */
    downloadUrl?: string;
}

/**
 * Check if a given object implements the DirectoryEntryDTO interface.
 */
export function instanceOfDirectoryEntryDTO(value: object): boolean {
    if (!('name' in value)) return false;
    if (!('path' in value)) return false;
    if (!('sha' in value)) return false;
    if (!('type' in value)) return false;
    return true;
}

export function DirectoryEntryDTOFromJSON(json: any): DirectoryEntryDTO {
    return DirectoryEntryDTOFromJSONTyped(json, false);
}

export function DirectoryEntryDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): DirectoryEntryDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'path': json['path'],
        'sha': json['sha'],
        'type': GitObjectTypeFromJSON(json['type']),
        'size': json['size'] == null ? undefined : json['size'],
        'downloadUrl': json['downloadUrl'] == null ? undefined : json['downloadUrl'],
    };
}

export function DirectoryEntryDTOToJSON(value?: DirectoryEntryDTO | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'path': value['path'],
        'sha': value['sha'],
        'type': GitObjectTypeToJSON(value['type']),
        'size': value['size'],
        'downloadUrl': value['downloadUrl'],
    };
}